Moved Inbox Panel component
This commit moves the Inbox Panel component
