+++ client/header/activity-panel/index.js
@@ -37,7 +37,9 @@ const HelpPanel = lazy( () =>
 );
 
 const InboxPanel = lazy( () =>
-	import( /* webpackChunkName: "activity-panels-inbox" */ './panels/inbox' )
+	import(
+		/* webpackChunkName: "activity-panels-inbox" */ '../../inbox-panel'
+	)
 );
 const OrdersPanel = lazy( () =>
 	import( /* webpackChunkName: "activity-panels-orders" */ './panels/orders' )
