+++ client/inbox-panel/placeholder.js
