+++ client/inbox-panel/style.scss
